requirements.txt
Flask==2.3.3
gunicorn==21.2.0
yt-dlp==2024.1.7
requests==2.31.0
groq==0.11.0
python-dotenv==1.0.0